requirements.txt
flask 
google-generativeai 
flask-cors 
langdetect 
requests 
cachetools